fix single quiz report

+++ src/main/java/com/letslive/letslearnbackend/entities/TopicQuizQuestion.java
@@ -8,9 +8,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 
 @Entity
@@ -42,7 +40,7 @@ public class TopicQuizQuestion {
     private String type;
 
     @JsonProperty("defaultMark")
-    private BigDecimal defaultMark;
+    private Double defaultMark;
 
     @JsonProperty("feedbackOfTrue")
     private String feedbackOfTrue;
